rename: hero 파일 section 폴더로 이동
